Added risk types
Risk types are added according to the mentioned conditions:

All upcoming (i.e. unsettled) bets from customers that win at an unusual
rate
should be highlighted as risky (it is up to you how this is shown)
Bets where the stake is more than 10 times higher than that customer’s
average
bet in their betting history should be highlighted as unusual (again, it
is up to you
how this is shown)
Bets where the stake is more than 30 times higher than that customer’s
average
bet in their betting history should be highlighted as highly unusual
Bets where the amount to be won is $1000 or more.